requirements.txt
requests
pandas
streamlit
dnspython
beautifulsoup4
validators
lxml
